added prov app controller to test connection
