--- src/main/java/com/batherphilippa/artapp/model/ArtworkOutput.java
@@ -1,13 +1,0 @@
-package com.batherphilippa.artapp.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ArtworkOutput {
-    private long id;
-    private String title;
-}
